Deduplicate the command test config helper
Extract the duplicated `test_command_config` setup from the `config` and
`pipes` test modules into a shared `#[cfg(test)] mod tests_support`, using
the default output and stream byte budgets. Both call sites import the
shared helper and drop their now-unused local setup imports.
