+++ src/stdlib/command/mod.rs
@@ -51,6 +51,8 @@ mod filters;
 mod pipes;
 mod quote;
 mod result;
+#[cfg(test)]
+mod tests_support;
 
 pub(super) use super::value_from_bytes;
 pub(crate) use config::CommandConfig;
